+++ src/Haya.Tool/Cli.fs
@@ -35,18 +35,32 @@ type BackstageArgs =
             | OutputPath _ -> "Path to output folder or file (default: ./haya.yaml)"
             | Format _ -> "Output format. json | yaml (default: yaml)"
 
+type DiagramArgs =
+    | [<MainCommand; ExactlyOnce; Last>] InputPath of input_path: string
+    | [<AltCommandLine("-o"); Unique>] OutputPath of output_path: string
+    | [<AltCommandLine("-t")>] DiagramType of DiagramType
+    | [<AltCommandLine("-f")>] DiagramFormat of DiagramFormat
+    interface IArgParserTemplate with
+        member s.Usage =
+            match s with
+            | InputPath _ -> "Path to solution or haya describe file"
+            | OutputPath _ -> "Path to output folder or file (default: ./c4-system-context.mermaid)"
+            | DiagramType _ -> "Output format. c4system (default: c4system)"
+            | DiagramFormat _ -> "Output format. mermaid (default: mermaid)"
+
 type CliArguments =
     | [<CliPrefix(CliPrefix.None)>] Crc of ParseResults<CrcArgs>
     | [<CliPrefix(CliPrefix.None)>] Describe of ParseResults<DescribeArgs>
     | [<CliPrefix(CliPrefix.None)>] Backstage of ParseResults<BackstageArgs>
+    | [<CliPrefix(CliPrefix.None)>] Diagram of ParseResults<DiagramArgs>
     interface IArgParserTemplate with
         member s.Usage =
             match s with
-            | Crc _ -> "Generate Components, Responsibilities, and Collaborator data"
+            | Crc _ -> "Generate Components, Responsibilities, and Collaborator documentation"
             | Describe _ -> "Generate Haya solution description"
             | Backstage _ -> "Generate Backstage catalog data"
+            | Diagram _ -> "Generate diagram files"
 
-    
 module Cli =
     open Haya.Core.Analysis
     open System
@@ -69,7 +83,7 @@ module Cli =
                    Format = format
                    CurrentDirectory = currentDir }
         else None
-        
+           
     let execDescribeAsync (cmd: DescribeCommand) = task {
         let pathToSln = cmd.PathToSln
         //todo: check if pathToSln is a sln or a describe file (json/yaml)
@@ -146,4 +160,35 @@ module Cli =
                 | Error e -> Error(1, e)
         else
             return Error(1, "Solution not found")
+    }
+    
+    let (|IsDiagramCommand|_|) (parserResults: ParseResults<CliArguments>) : DiagramCommand option =
+        if parserResults.Contains(CliArguments.Diagram) then
+            let currentDir = Environment.CurrentDirectory
+            let pr = parserResults.GetResult(CliArguments.Diagram)
+            let sln = pr.GetResult(DiagramArgs.InputPath)
+            let outputPath: string = pr.TryGetResult(DiagramArgs.OutputPath) |> Option.defaultValue "./c4-system-context.mermaid"
+            let diagramType = pr.GetResult(DiagramArgs.DiagramType, defaultValue = DiagramType.C4System)
+            let diagramFormat = pr.GetResult(DiagramArgs.DiagramFormat, defaultValue = DiagramFormat.Mermaid)
+            Some { PathToSln = sln
+                   OutputPath = outputPath
+                   DiagramType = diagramType
+                   DiagramFormat = diagramFormat
+                   CurrentDirectory = currentDir }
+        else None
+        
+    let execDiagramAsync (cmd: DiagramCommand) = task {
+        let pathToSln = cmd.PathToSln
+        if IO.fileExists pathToSln then
+            let! sln = pathToSln |> Roslyn.openSolution
+            let! descriptors =
+                sln |> Descriptor.getDescriptors cmd.CurrentDirectory Descriptor.attributeNames
+            return
+                Diagram.generateDiagram cmd descriptors
+                |> IO.write cmd.OutputPath
+                |> function
+                | Ok _ -> Ok(0, $"Diagram successfully generated from {pathToSln}")
+                | Error e -> Error(1, e)
+        else
+            return Error(1, "Solution not found")    
     }