+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `version` field for the internal spec.
 - YAML support as an output format.
 - `Owner` on CollaboratorAttribute.
+- `diagram` command for outputting diagrams. Currently only mermaid support.
 
 ### Changed